requirements.txt
sentence-transformers==2.2.2
Flask==2.0.1
pandas==1.3.0
scikit-learn==0.24.2
numpy==1.21.0
google-cloud-storage==2.9.0
